fix(withdrawal): 提现打款审查问题批量修复（C1-C2/H1-H3/M1-M8/L4-L5）
致命修复：
- C1: 单笔approve不再接受手动txHash（安全绕过风险），统一走approveWithTransfer
  两阶段链上打款，与batchApprove共用同一条路径
- C2: Phase3 markApprovedAfterBroadcast失败时记录CRITICAL日志+txHash供人工对账

高危修复：
- H1: EthereumClient nonce管理——进程内AtomicLong计数器按地址隔离，
  nonce冲突时自动resync+重试最多3次，解决批量打款nonce冲突
- H2: batchApprove改为continue尽力而为模式，某条失败后继续处理剩余
- H3: WalletService.bindOrReplace加userActionLocks，防止与提现create并发
  时wallet被覆盖导致网络/地址错配

中危修复：
- M1: create()校验withdrawable>0，防止feePercent配置过高导致全额被扣
- M2: EthereumClient余额查询失败抛异常（与TronClient一致），不再静默返回0
- M3: batchPreview余额不足预警（按链分别检查）
- M4: gas price估算失败记录warn日志而非静默吞掉
- M5: ERC20 transfer ABI返回类型Uint8改Bool
- M6: Android费用计算用Long防Int溢出
- M7: admin-web doBatchApprove在finally块清私钥（失败路径也清）
- M8: admin-web el-dialog @close绑定closeBatchDialog（ESC/X关闭也清理）

低危修复：
- L4: WithdrawalRequest.create setScale改HALF_UP避免配置小数过多时抛异常
- L5: admin-web批量结果PENDING项标记info色（非danger红色）

331个后端测试通过，Android编译通过，admin-web类型检查通过。

+++ android-app/app/src/main/java/com/reelshort/app/ui/format/UiFormats.kt
@@ -271,7 +271,8 @@ internal fun withdrawalConversionLines(
             "1 USD = ${cnyPerUsd.toUiDecimal()} CNY",
     )
     if (pointAmount > 0) {
-        val fee = if (summary.feePercent > 0) (pointAmount * summary.feePercent + 99) / 100 else 0
+        // M6: 用 Long 防溢出（pointAmount * feePercent 在大额时可能超过 Int.MAX_VALUE）
+        val fee = if (summary.feePercent > 0) ((pointAmount.toLong() * summary.feePercent + 99) / 100).toInt() else 0
         val withdrawable = pointAmount - fee
         val points = withdrawable.toBigDecimal()
         val cnyAmount = points.multiply(cnyPerPoint)
